move openssl asn1 encode functions to a new module
